% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileStore.R
\name{.blob_get}
\alias{.blob_get}
\title{Read a blob (checks memory first, then disk)}
\usage{
.blob_get(key, ss)
}
\arguments{
\item{key}{Character. Blob key}

\item{ss}{Session environment}
}
\value{
Character string or NULL
}
\description{
Read a blob (checks memory first, then disk)
}
\keyword{internal}
