% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileStore.R
\name{.blob_nuke}
\alias{.blob_nuke}
\title{Delete all blobs (memory + disk)}
\usage{
.blob_nuke(ss)
}
\arguments{
\item{ss}{Session environment}
}
\description{
Delete all blobs (memory + disk)
}
\keyword{internal}
