% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileStore.R
\name{.blob_put}
\alias{.blob_put}
\title{Write a blob (adaptive: memory for small, disk for large)}
\usage{
.blob_put(key, value, ss)
}
\arguments{
\item{key}{Character. Blob key}

\item{value}{Character. Blob data (base64 string)}

\item{ss}{Session environment}
}
\description{
Write a blob (adaptive: memory for small, disk for large)
}
\keyword{internal}
