% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileStore.R
\name{.blob_read}
\alias{.blob_read}
\title{Read a blob without consuming it}
\usage{
.blob_read(key, ss)
}
\arguments{
\item{key}{Character. Blob key}

\item{ss}{Session environment}
}
\value{
Character string or NULL
}
\description{
Read a blob without consuming it
}
\keyword{internal}
