% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileStore.R
\name{.blob_snapshot}
\alias{.blob_snapshot}
\title{Read ALL blobs into a named list (memory + disk merged)}
\usage{
.blob_snapshot(ss)
}
\arguments{
\item{ss}{Session environment}
}
\value{
Named list of blob values
}
\description{
Read ALL blobs into a named list (memory + disk merged)
}
\keyword{internal}
