% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhe_utils.R
\name{.callMheTool}
\alias{.callMheTool}
\title{Call mhe-tool with JSON input}
\usage{
.callMheTool(command, input_data)
}
\arguments{
\item{command}{The mhe-tool subcommand to run}

\item{input_data}{List that will be converted to JSON input}
}
\value{
Parsed JSON output from mhe-tool
}
\description{
Call mhe-tool with JSON input
}
\keyword{internal}
