% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcUtils.R
\name{.callMpcTool}
\alias{.callMpcTool}
\title{Call dsvert-mpc with JSON input}
\usage{
.callMpcTool(command, input_data)
}
\arguments{
\item{command}{The dsvert-mpc subcommand to run}

\item{input_data}{List that will be converted to JSON input}
}
\value{
Parsed JSON output from dsvert-mpc
}
\description{
Call dsvert-mpc with JSON input
}
\keyword{internal}
