% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcUtils.R
\name{.dsvert_mpc_encode_json}
\alias{.dsvert_mpc_encode_json}
\title{Call dsvert-mpc with JSON input}
\usage{
.dsvert_mpc_encode_json(input_data)
}
\arguments{
\item{input_data}{List that will be converted to JSON input}
}
\value{
Parsed JSON output from dsvert-mpc
}
\description{
Call dsvert-mpc with JSON input
}
\keyword{internal}
