% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsiRelay.R
\name{.dsvert_relay_close}
\alias{.dsvert_relay_close}
\title{Remove only a relay-created private spool directory.}
\usage{
.dsvert_relay_close(ss)
}
\description{
Remove only a relay-created private spool directory.
}
\keyword{internal}
