% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsiRelay.R
\name{.dsvert_relay_gc}
\alias{.dsvert_relay_gc}
\title{Expire a relay session after a public, fixed period without activity.}
\usage{
.dsvert_relay_gc(ss, now = .dsvert_relay_now())
}
\description{
Expire a relay session after a public, fixed period without activity.
}
\keyword{internal}
