% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcUtils.R
\name{.dsvert_validate_logical_peer_name}
\alias{.dsvert_validate_logical_peer_name}
\title{Validate a consortium logical peer name}
\usage{
.dsvert_validate_logical_peer_name(name)
}
\description{
Validate a consortium logical peer name
}
\keyword{internal}
