% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhe_utils.R
\name{.findMheTool}
\alias{.findMheTool}
\title{MHE Utility Functions}
\usage{
.findMheTool()
}
\value{
Path to the mhe-tool binary
}
\description{
Internal functions for calling the mhe-tool binary.
}
\keyword{Find}
\keyword{binary}
\keyword{internal}
\keyword{mhe-tool}
\keyword{the}
