% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcUtils.R
\name{.get_identity_keypair}
\alias{.get_identity_keypair}
\title{Derive Ed25519 identity keypair from seed}
\usage{
.get_identity_keypair()
}
\value{
List with identity_pk and identity_sk (standard base64).
}
\description{
Derive Ed25519 identity keypair from seed
}
\keyword{internal}
