% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcUtils.R
\name{.get_trusted_peers}
\alias{.get_trusted_peers}
\title{Get the name-bound map of trusted peer identity PKs}
\usage{
.get_trusted_peers()
}
\value{
Named character vector mapping logical peer names to base64 PKs.
}
\description{
Get the name-bound map of trusted peer identity PKs
}
\keyword{internal}
