% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileStore.R
\name{.key_delete}
\alias{.key_delete}
\title{Delete a persistent key (memory + disk with secure overwrite)}
\usage{
.key_delete(name, ss)
}
\arguments{
\item{name}{Character. Key name}

\item{ss}{Session environment}
}
\description{
Delete a persistent key (memory + disk with secure overwrite)
}
\keyword{internal}
