% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileStore.R
\name{.key_exists}
\alias{.key_exists}
\title{Check if a persistent key exists (memory or disk)}
\usage{
.key_exists(name, ss)
}
\arguments{
\item{name}{Character. Key name}

\item{ss}{Session environment}
}
\value{
Logical
}
\description{
Check if a persistent key exists (memory or disk)
}
\keyword{internal}
