% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileStore.R
\name{.key_get}
\alias{.key_get}
\title{Read a persistent key (memory first, then disk)}
\usage{
.key_get(name, ss)
}
\arguments{
\item{name}{Character. Key name}

\item{ss}{Session environment}
}
\value{
Character string (or vector for multi-line), or NULL
}
\description{
Read a persistent key (memory first, then disk)
}
\keyword{internal}
