% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileStore.R
\name{.key_put}
\alias{.key_put}
\title{Write a persistent key to session memory}
\usage{
.key_put(name, value, ss)
}
\arguments{
\item{name}{Character. Key name (e.g., "cpk", "secret_key")}

\item{value}{Character. Key data (single string or character vector)}

\item{ss}{Session environment}
}
\description{
Write a persistent key to session memory
}
\keyword{internal}
