% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileStore.R
\name{.session_dir_cleanup}
\alias{.session_dir_cleanup}
\title{Clean up all session storage (memory + disk with secure deletion)}
\usage{
.session_dir_cleanup(ss)
}
\arguments{
\item{ss}{Session environment}
}
\description{
Clean up all session storage (memory + disk with secure deletion)
}
\keyword{internal}
