% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcUtils.R
\name{.sign_transport_pk}
\alias{.sign_transport_pk}
\title{Sign a transport PK with the identity SK}
\usage{
.sign_transport_pk(transport_pk_b64, identity_sk_b64)
}
\value{
Character. Standard base64 signature.
}
\description{
Sign a transport PK with the identity SK
}
\keyword{internal}
