% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcUtils.R
\name{.verify_peer_identity}
\alias{.verify_peer_identity}
\title{Verify a peer's transport PK signature}
\usage{
.verify_peer_identity(transport_pk_b64, identity_pk_b64, signature_b64)
}
\value{
Logical.
}
\description{
Verify a peer's transport PK signature
}
\keyword{internal}
