% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsvertTestHelpersDS.R
\name{dsvertAddQuartileColumnDS}
\alias{dsvertAddQuartileColumnDS}
\title{Append an age-quartile factor column (test helper)}
\usage{
dsvertAddQuartileColumnDS(data_name, column = "age", output_column = "age_q")
}
\arguments{
\item{data_name}{Character. Name of the data frame symbol on the server.}

\item{column}{Character. Name of an existing column to operate on.}

\item{output_column}{Character. Name of the new column to add to the data frame.}
}
\description{
Compute quartile boundaries of the named column
(defaulting to \code{age}) and attach a factor column with levels
Q1..Q4. Used to validate \code{ds.vertChisqCross}.
}
\keyword{internal}
