% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsvertTestHelpersDS.R
\name{dsvertCopyDfDS}
\alias{dsvertCopyDfDS}
\title{Copy a data frame to a new name (test helper)}
\usage{
dsvertCopyDfDS(data_name, output_name)
}
\arguments{
\item{data_name}{Character. Name of the data frame symbol on the server.}

\item{output_name}{Character. Name to bind the output object to in the server-side environment.}
}
\description{
Copy a data frame to a new name (test helper)
}
