% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jointDPVectorAllocatorDS.R
\name{dsvertJointDPVectorAllocationAuthorizeDS}
\alias{dsvertJointDPVectorAllocationAuthorizeDS}
\title{Authorize the biomedical capsule allocation (AGGREGATE)}
\usage{
dsvertJointDPVectorAllocationAuthorizeDS(first_commit_json, second_commit_json)
}
\arguments{
\item{first_commit_json, second_commit_json}{Signed commit receipts.}
}
\value{
A signed authorization receipt.
}
\description{
Authorize the biomedical capsule allocation (AGGREGATE)
}
\keyword{internal}
