% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbFullRegShareDS.R
\name{dsvertNBYThetaShareReceiveDS}
\alias{dsvertNBYThetaShareReceiveDS}
\title{NL-side: receive (y + theta) share blob + store under canonical key}
\usage{
dsvertNBYThetaShareReceiveDS(peer_yt_share_blob_key, session_id = NULL)
}
\arguments{
\item{peer_yt_share_blob_key}{Character.}

\item{session_id}{Character.}
}
\description{
Decrypts the transport blob from
\code{dsvertNBYThetaShareDS} and stores in
\code{ss$k2_nb_yt_share_fp}.
}
