% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactGCChisqGuardDS.R
\name{exactGCChisqGuardFinalizeDS}
\alias{exactGCChisqGuardFinalizeDS}
\title{Seal the aggregate cross-contingency guard share to its peer (AGGREGATE)}
\usage{
exactGCChisqGuardFinalizeDS(manifest_handle, operation_id, session_id)
}
\description{
Seal the aggregate cross-contingency guard share to its peer (AGGREGATE)
}
\keyword{internal}
