% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactGCChisqGuardDS.R
\name{exactGCChisqGuardPrepareDS}
\alias{exactGCChisqGuardPrepareDS}
\title{Prepare the purpose-bound cross-contingency count guard (AGGREGATE)}
\usage{
exactGCChisqGuardPrepareDS(session_id)
}
\description{
Prepare the purpose-bound cross-contingency count guard (AGGREGATE)
}
\keyword{internal}
