% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactGCChisqGuardDS.R
\name{exactGCChisqGuardStartDS}
\alias{exactGCChisqGuardStartDS}
\title{Start the purpose-bound cross-contingency count guard (AGGREGATE)}
\usage{
exactGCChisqGuardStartDS(manifest_handle, operation_id, session_id)
}
\description{
Start the purpose-bound cross-contingency count guard (AGGREGATE)
}
\keyword{internal}
