% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactGCTransportDS.R
\name{exactGCTransportInitDS}
\alias{exactGCTransportInitDS}
\title{Initialize an idempotent exact-GC peer transport (AGGREGATE)}
\usage{
exactGCTransportInitDS(session_id)
}
\arguments{
\item{session_id}{Active exact-GC session identifier.}
}
\description{
Initialize an idempotent exact-GC peer transport (AGGREGATE)
}
