% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_corr.R
\name{get_corr}
\alias{get_corr}
\title{Get Correlation Matrix from a single dataset using SVD components}
\usage{
get_corr(data)
}
\arguments{
\item{data}{any matrix containing numerical variables}
}
\value{
correlation matrix
}
\description{
Get Correlation Matrix from a single dataset using SVD components
}
\examples{
get_corr(any_data)

}
