% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmRing63DS.R
\name{glmRing63CorSetZeroYDS}
\alias{glmRing63CorSetZeroYDS}
\title{Set y_share to zeros (for correlation: no response variable)}
\usage{
glmRing63CorSetZeroYDS(session_id = NULL)
}
\arguments{
\item{session_id}{Character or NULL.}
}
\value{
List with status.
}
\description{
Set y_share to zeros (for correlation: no response variable)
}
