% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_text.R
\name{hash_text}
\alias{hash_text}
\title{Hash list of text or IDs}
\usage{
hash_text(strings, method = "additive")
}
\arguments{
\item{strings}{list of string elements to be turned into hashes}

\item{method}{additive or polynomial methods currently implemented}
}
\value{
a vector of the original list size containing the hash values
}
\description{
Hash list of text or IDs
}
\examples{
hash_text(c("one", "two", "three"))
hash_text(c("one", "two", "three"), "additive")
hash_text(c("one", "two", "three"), method = "polynomial")

}
