% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k2WideSplineFullDS.R
\name{k2-wide-spline}
\alias{k2-wide-spline}
\title{K=2 DCF Wide Spline (4-Phase Sigmoid/Exp)}
\description{
Piecewise-linear sigmoid/exp evaluation using Distributed
Comparison Functions (DCF) and Beaver triples. All computation in
Ring63 fixed-point.
}
