% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactGCChisqGuardDS.R
\name{k2ChisqCrossAccumulateCountDS}
\alias{k2ChisqCrossAccumulateCountDS}
\title{Accumulate one fixed cross-contingency cell share (AGGREGATE)}
\usage{
k2ChisqCrossAccumulateCountDS(cell_index, total_cells, session_id)
}
\description{
Accumulate one fixed cross-contingency cell share (AGGREGATE)
}
\keyword{internal}
