% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k2InputSharingDS.R
\name{k2GradientR1DS}
\alias{k2GradientR1DS}
\title{Gradient round 1: compute (X-A, r-B) in Ring63}
\usage{
k2GradientR1DS(peer_pk, session_id = NULL)
}
\description{
Gradient round 1: compute (X-A, r-B) in Ring63
}
