% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otBeaverDS.R
\name{k2OtMulSenderSetupDS}
\alias{k2OtMulSenderSetupDS}
\title{Start an OT multiplication sender transcript}
\usage{
k2OtMulSenderSetupDS(ot_key, session_id = NULL)
}
\arguments{
\item{ot_key}{Session key prefix for one cross-term direction.}

\item{session_id}{MPC session id.}
}
\value{
Public sender setup to relay to the receiver.
}
\description{
Start an OT multiplication sender transcript
}
\keyword{internal}
