% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k2InputSharingDS.R
\name{k2ReceiveShareDS}
\alias{k2ReceiveShareDS}
\title{Receive peer's shared data (FixedPoint)}
\usage{
k2ReceiveShareDS(peer_p = NULL, peer_name, session_id = NULL)
}
\arguments{
\item{peer_p}{Integer. Number of features held by the peer (sets share length).}

\item{peer_name}{Authenticated logical name of the expected producer.}

\item{session_id}{Character. Active MPC session identifier.}
}
\description{
Receive peer's shared data (FixedPoint)
}
