% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mheFullProtocol.R
\name{mheEncryptLocalDS}
\alias{mheEncryptLocalDS}
\title{Encrypt local data columns using stored CPK}
\usage{
mheEncryptLocalDS(data_name, variables)
}
\arguments{
\item{data_name}{Character. Name of data frame}

\item{variables}{Character vector. Variables to encrypt}
}
\value{
List with encrypted_columns
}
\description{
Encrypt local data columns using stored CPK
}
