% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mheFullProtocol.R
\name{mhePartialDecryptDS}
\alias{mhePartialDecryptDS}
\title{Compute partial decryption using stored secret key and stored ciphertext chunks}
\usage{
mhePartialDecryptDS(n_chunks)
}
\arguments{
\item{n_chunks}{Integer. Number of stored ciphertext chunks}
}
\value{
List with decryption_share (chunked as a character vector)
}
\description{
Compute partial decryption using stored secret key and stored ciphertext chunks
}
