% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mheFullProtocol.R
\name{mheStoreEncChunkDS}
\alias{mheStoreEncChunkDS}
\title{Store a chunk of an encrypted column (for transferring large ciphertexts)}
\usage{
mheStoreEncChunkDS(col_index, chunk_index, chunk)
}
\arguments{
\item{col_index}{Integer. Column index (1-based)}

\item{chunk_index}{Integer. Chunk index (1-based)}

\item{chunk}{Character. Base64url-encoded chunk of ciphertext}
}
\value{
TRUE on success
}
\description{
Store a chunk of an encrypted column (for transferring large ciphertexts)
}
