% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderTableDS.R
\name{reorderTableDS}
\alias{reorderTableDS}
\title{Serverside function for ordering a dataset according to a list of hashes of the same id variable}
\usage{
reorderTableDS(table_name, id_var, hash_list, new_table_name = "orderedTable")
}
\arguments{
\item{table_name}{The dataframe that needs to be ordered}

\item{id_var}{The id variable to be hashed (if this is different from the one used in hash_list we end up with an ordered df with all NULL VALUES)}

\item{hash_list}{List of hashes to be used for ordering}

\item{new_table_name}{Name of new object to be created that is in the same order as the hash_list}
}
\value{
Should assign a new object only
}
\description{
Serverside function for ordering a dataset according to a list of hashes of the same id variable
}
\examples{
reorderTableDS(D,"id",all_hashes, "orderedTable")
}
