% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertCorDS.R
\name{vertCorDS}
\alias{vertCorDS}
\title{Serverside function used for vertCor}
\usage{
vertCorDS(variable_names = NULL)
}
\arguments{
\item{variable_names}{variables to be used in correlation matrix}
}
\value{
the combined matrix for the intermidiary step of the block SVD algorithm
}
\description{
Serverside function used for vertCor
}
\examples{
TODO
}
