% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vertical_corr.R
\name{vertical_corr}
\alias{vertical_corr}
\title{Get Correlation Matrix for two Vertically Partitioned datasets}
\usage{
vertical_corr(data1, data2)
}
\arguments{
\item{data1}{Partition 1}

\item{data2}{Partition 2 w/ assumption that the two datasets are already alligned}
}
\value{
correlation matrix obtained via block SVD results of the two partitions
}
\description{
Get Correlation Matrix for two Vertically Partitioned datasets
}
\examples{
vertical_corr(Partition1, Partition2)

}
