% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileStore.R
\name{.blob_consume}
\alias{.blob_consume}
\title{Read a blob and delete it (one-shot consumption)}
\usage{
.blob_consume(key, ss)
}
\arguments{
\item{key}{Character. Blob key}

\item{ss}{Session environment}
}
\value{
Character string or NULL
}
\description{
Read a blob and delete it (one-shot consumption)
}
\keyword{internal}
