% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcUtils.R
\name{.get_identity_seed}
\alias{.get_identity_seed}
\title{Get the identity seed (R option or filesystem)}
\usage{
.get_identity_seed()
}
\value{
Character. Base64-encoded seed.
}
\description{
Get the identity seed (R option or filesystem)
}
\keyword{internal}
