% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcUtils.R
\name{.get_trusted_peers}
\alias{.get_trusted_peers}
\title{Get the set of trusted peer identity PKs}
\usage{
.get_trusted_peers()
}
\value{
Character vector of base64 PKs, or NULL if enforcement disabled.
}
\description{
Get the set of trusted peer identity PKs
}
\keyword{internal}
