% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\encoding{UTF-8}
\name{dsVert-package}
\alias{dsVert}
\alias{dsVert-package}
\title{Package Load Hook}
\description{
Auto-generates Ed25519 identity seed on first load if none exists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/isglobal-brge/dsVert}
  \item Report bugs at \url{https://github.com/isglobal-brge/dsVert/issues}
}

}
\author{
\strong{Maintainer}: David Sarrat González \email{david.sarrat@isglobal.org}

Other contributors:
\itemize{
  \item Miron Banjac \email{miron.banjac@isglobal.org} [contributor]
  \item Juan R González \email{juanr.gonzalez@isglobal.org} [contributor]
}

}
\keyword{internal}
