% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsvertTestHelpersDS.R
\name{dsvertCopyDfDS}
\alias{dsvertCopyDfDS}
\title{Copy a data frame to a new name (test helper)}
\usage{
dsvertCopyDfDS(data_name, output_name)
}
\description{
Copy a data frame to a new name (test helper)
}
