% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashIdDS.R
\name{hashIdDS}
\alias{hashIdDS}
\title{Serverside function used to has an Id variable of a dataframe, taking full advantage of digest library}
\usage{
hashIdDS(data_name, id_variable, algo = "sha256")
}
\arguments{
\item{data_name}{The name of the dataframe on the server}

\item{id_variable}{Variable to be hashed}

\item{algo}{Which digest algorithm to use, it is set to sha256 is a base}
}
\value{
vector of hashes
}
\description{
Serverside function used to has an Id variable of a dataframe, taking full advantage of digest library
}
\examples{
hashIdDS(D, "id", "sha256")
}
