% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxDS.R
\name{k2CoxForwardCumsumGDS}
\alias{k2CoxForwardCumsumGDS}
\title{Compute the forward cumsum G_j = sum_{i<=j, delta=1} recip\link{i}}
\usage{
k2CoxForwardCumsumGDS(session_id = NULL)
}
\arguments{
\item{session_id}{GLM session id.}
}
\value{
list(G_length)
}
\description{
Compute the forward cumsum G_j = sum_{i<=j, delta=1} recip\link{i}
}
