% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k2InputSharingDS.R
\name{k2GradientR2DS}
\alias{k2GradientR2DS}
\title{Gradient round 2: compute gradient share from Beaver formula}
\usage{
k2GradientR2DS(party_id = 0L, session_id = NULL)
}
\description{
Gradient round 2: compute gradient share from Beaver formula
}
