% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otBeaverDS.R
\name{k2IknpBaseSenderFinalizeDS}
\alias{k2IknpBaseSenderFinalizeDS}
\title{Finalise IKNP base-OT sender state}
\usage{
k2IknpBaseSenderFinalizeDS(
  ciphertexts = NULL,
  ciphertexts_blob_key = NULL,
  iknp_key,
  session_id = NULL
)
}
\arguments{
\item{ciphertexts}{Public base ciphertexts.}

\item{ciphertexts_blob_key}{Optional blob key containing ciphertexts.}

\item{iknp_key}{Session key prefix.}

\item{session_id}{MPC session id.}
}
\value{
list(stored = TRUE).
}
\description{
Finalise IKNP base-OT sender state
}
