% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k2InputSharingDS.R
\name{k2ReceiveShareDS}
\alias{k2ReceiveShareDS}
\title{Receive peer's shared data (FixedPoint)}
\usage{
k2ReceiveShareDS(peer_p = NULL, session_id = NULL)
}
\description{
Receive peer's shared data (FixedPoint)
}
