% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k2InputSharingDS.R
\name{k2ShareInputDS}
\alias{k2ShareInputDS}
\title{Share local data with peer (FixedPoint shares)}
\usage{
k2ShareInputDS(data_name, x_vars, y_var = NULL, peer_pk, session_id = NULL)
}
\description{
Share local data with peer (FixedPoint shares)
}
