% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k2InputSharingDS.R
\name{k2StoreGradTripleDS}
\alias{k2StoreGradTripleDS}
\title{Store gradient Beaver triple (Ring63 FP format)}
\usage{
k2StoreGradTripleDS(session_id = NULL)
}
\description{
Store gradient Beaver triple (Ring63 FP format)
}
