% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mheFullProtocol.R
\name{mheAssembleEncColumnDS}
\alias{mheAssembleEncColumnDS}
\title{Assemble stored chunks into a complete encrypted column}
\usage{
mheAssembleEncColumnDS(col_index, n_chunks)
}
\arguments{
\item{col_index}{Integer. Column index (1-based)}

\item{n_chunks}{Integer. Total number of chunks}
}
\value{
TRUE on success
}
\description{
Assemble stored chunks into a complete encrypted column
}
