% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mheFullProtocol.R
\name{mheCombineDS}
\alias{mheCombineDS}
\title{Combine public key shares into collective public key}
\usage{
mheCombineDS(public_key_shares, crp, num_obs = 100, log_n = 12, log_scale = 40)
}
\arguments{
\item{public_key_shares}{Character vector. Public key shares from all servers}

\item{crp}{Character. CRP from party 0}

\item{num_obs}{Integer. Number of observations}

\item{log_n}{Integer. Ring dimension}

\item{log_scale}{Integer. Scale parameter}
}
\value{
List with collective_public_key
}
\description{
Combine public key shares into collective public key
}
