% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmEncryptedDS.R
\name{mheEncryptRawDS}
\alias{mheEncryptRawDS}
\title{Encrypt raw response variable under CPK (label server only)}
\usage{
mheEncryptRawDS(data_name, y_var)
}
\arguments{
\item{data_name}{Character. Name of data frame}

\item{y_var}{Character. Name of response variable}
}
\value{
List with encrypted_y (base64url) and num_obs
}
\description{
Encrypt raw response variable under CPK (label server only)
}
