% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mheFullProtocol.R
\name{mheGetObsDS}
\alias{mheGetObsDS}
\title{Get number of observations for a variable}
\usage{
mheGetObsDS(data_name, variables)
}
\arguments{
\item{data_name}{Character. Name of data frame}

\item{variables}{Character vector. Variables to check}
}
\value{
Integer. Number of complete observations
}
\description{
Get number of observations for a variable
}
