% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mheFullProtocol.R
\name{mheStoreCPKDS}
\alias{mheStoreCPKDS}
\title{Store collective public key received from combine step}
\usage{
mheStoreCPKDS(cpk, galois_keys = NULL, relin_key = NULL)
}
\arguments{
\item{cpk}{Character. Collective public key (base64url)}

\item{galois_keys}{Character vector. Galois keys (base64url)}

\item{relin_key}{Character. Relinearization key (base64url)}
}
\value{
TRUE on success
}
\description{
Store collective public key received from combine step
}
