% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mheFullProtocol.R
\name{mheStoreCTChunkDS}
\alias{mheStoreCTChunkDS}
\title{Store a chunk of a ciphertext for partial decryption}
\usage{
mheStoreCTChunkDS(chunk_index, chunk)
}
\arguments{
\item{chunk_index}{Integer. Chunk index (1-based)}

\item{chunk}{Character. Base64url-encoded chunk}
}
\value{
TRUE on success
}
\description{
Store a chunk of a ciphertext for partial decryption
}
