% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmEncryptedDS.R
\name{mheStoreEncYDS}
\alias{mheStoreEncYDS}
\title{Store encrypted response variable (non-label servers)}
\usage{
mheStoreEncYDS(enc_y)
}
\arguments{
\item{enc_y}{Character. Encrypted y ciphertext (base64url, may be chunked)}
}
\value{
TRUE on success
}
\description{
Store encrypted response variable (non-label servers)
}
