% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpcUtils.R
\name{mpcAvailable}
\alias{mpcAvailable}
\title{Check if MPC binary is available}
\usage{
mpcAvailable()
}
\value{
TRUE if dsvert-mpc binary is available, FALSE otherwise
}
\description{
Check if MPC binary is available
}
