% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmRing63DS.R
\name{mpcStoreBlobDS}
\alias{mpcStoreBlobDS}
\title{Store a blob on server (adaptive chunking support)}
\usage{
mpcStoreBlobDS(key, chunk, chunk_index = 1L, n_chunks = 1L, session_id = NULL)
}
\arguments{
\item{key}{Character. Blob key.}

\item{chunk}{Character. Blob data (or chunk if multi-part).}

\item{chunk_index}{Integer. Chunk index (1-based).}

\item{n_chunks}{Integer. Total chunks.}

\item{session_id}{Character or NULL.}
}
\value{
TRUE on success.
}
\description{
Store a blob on server (adaptive chunking support)
}
